Create C project
This project The project was to connect a raspberry pi to a temperature sensor and an LCD panel, then collect data from the sensor, track it onto an online database, and then find specific values of the data. For example finding the average, maximum, or minimum temperature of the data. My program took data and every 10 seconds from 5-19-2022 to 5-22-2022.